Create Players class
