+++ bin/main.rb
@@ -2,22 +2,8 @@
 
 # rubocop: disable Metrics/CyclomaticComplexity, Metrics/MethodLength, Metrics/PerceivedComplexity, Metrics/AbcSize, Metrics/BlockNesting
 
+require_relative '../lib/players.rb'
 require_relative '../lib/logic'
-
-def player_assignment
-  puts 'Welcome to this tic-tac-toe game'
-  puts 'Please, put in your name Player 1'
-  player_one = gets.chomp
-  sleep 0.5
-  puts "Player 1 your marker is 'X'"
-  puts ''
-  puts 'Please, put in your name Player 2'
-  player_two = gets.chomp
-  sleep 0.5
-  puts "Player 2 your marker is 'O'"
-  puts ''
-  [player_one, player_two]
-end
 
 def board(arr)
   puts '+---+---+---+'
@@ -31,7 +17,7 @@ end
 
 def play
   cell = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-  players = player_assignment
+  players = Players.new.player_assignment
   board(cell)
   i = 0
   game_on = true
